requirements.txt
flask
requests
sympy
beautifulsoup4
openai
python-dotenv